archive/cloud-functions/intervene-trigger/requirements.txt
# Flask is used to create the HTTP server
Flask==2.3.2
Flask-Cors==4.0.0

# Google Cloud Pub/Sub client library
google-cloud-pubsub==2.13.4

# Firebase Admin SDK for verifying auth tokens
firebase-admin==6.1.0

# PDF processing security fix
pypdf>=6.0.0
